Fixed build errors, changed ints to longs for id

+++ assessments/src/main/java/com/revature/caliber/assessments/data/Facade.java
@@ -97,7 +97,7 @@ public interface Facade {
     /**
      * Returns a grade object given a specific gradeId
      */
-    Grade getGradeByGradeId(int gradeId);
+    Grade getGradeByGradeId(long gradeId);
 
     /**
      * Returns a list of grades of a specific trainee based on traineeId as an input
@@ -111,7 +111,7 @@ public interface Facade {
      *
      * @param assessmentId
      */
-    List<Grade> getGradesByAssesessment(int assessmentId);
+    List<Grade> getGradesByAssesessment(long assessmentId);
 
     //Insert
